✨ feat[#1]: DB 구조 변경에 따라 mycomputerdto 수정사항 반영
